requirements.txt
azure-ai-vision==0.15.1b1
uvicorn~=0.23.2
fastapi~=0.104.1
requests~=2.31.0
httpx~=0.25.1
openai~=1.2.4
python-dotenv~=1.0.0